+++ packages/sit-onyx/src/components/OnyxTextarea/OnyxTextarea.vue
@@ -1,12 +1,10 @@
 <script lang="ts" setup>
 import { computed } from "vue";
 import { useDensity } from "../../composables/density";
-import { useRequired } from "../../composables/required";
 import { useCustomValidity } from "../../composables/useCustomValidity";
+import OnyxFormElement from "../OnyxFormElement/OnyxFormElement.vue";
 import OnyxSkeleton from "../OnyxSkeleton/OnyxSkeleton.vue";
-import OnyxInfoTooltip from "../OnyxInfoTooltip/OnyxInfoTooltip.vue";
 import type { OnyxTextareaProps } from "./types";
-import { injectI18n } from "../../i18n";
 
 const props = withDefaults(defineProps<OnyxTextareaProps>(), {
   modelValue: "",
@@ -43,7 +41,6 @@ const emit = defineEmits<{
 
 const { vCustomValidity, errorMessages } = useCustomValidity({ props, emit });
 
-const { requiredMarkerClass, requiredTypeClass } = useRequired(props);
 const { densityClass } = useDensity(props);
 
 /**
@@ -58,9 +55,6 @@ const handleChange = (event: Event) => {
   const inputValue = (event.target as HTMLInputElement).value;
   emit("change", inputValue);
 };
-
-const { t } = injectI18n();
-const shouldShowCounter = computed(() => props.withCounter && props.maxlength);
 
 /**
  * Current CSS variables for the autosize min/max height.
@@ -92,28 +86,8 @@ const handleInput = (event: Event) => {
     <OnyxSkeleton class="onyx-textarea-skeleton__input" />
   </div>
 
-  <div
-    v-else
-    :class="['onyx-textarea', requiredTypeClass, densityClass]"
-    :style="autosizeMinMaxStyles"
-  >
-    <label>
-      <div
-        v-if="!props.hideLabel"
-        class="onyx-textarea__label onyx-text--small"
-        :class="[!props.required ? requiredMarkerClass : undefined]"
-      >
-        <div class="onyx-textarea__header">
-          <span class="onyx-truncation-ellipsis">{{ props.label }}</span>
-          <span
-            v-if="props.required"
-            :class="[props.required ? requiredMarkerClass : undefined]"
-          ></span>
-          <OnyxInfoTooltip v-if="props.labelTooltip" :text="props.labelTooltip" />
-          <span v-if="!props.required" class="onyx-textarea__optional">{{ t("optional") }}</span>
-        </div>
-      </div>
-
+  <div v-else :class="['onyx-textarea', densityClass]" :style="autosizeMinMaxStyles">
+    <OnyxFormElement v-bind="props" :error-messages="errorMessages">
       <div class="onyx-textarea__wrapper" :data-autosize-value="value">
         <!-- eslint-disable vuejs-accessibility/no-autofocus -
          We want to provide the flexibility to have the autofocus property.
@@ -142,35 +116,7 @@ const handleInput = (event: Event) => {
         ></textarea>
         <!-- eslint-enable vuejs-accessibility/no-autofocus -->
       </div>
-    </label>
-
-    <div
-      v-if="props.message || errorMessages?.shortMessage || shouldShowCounter"
-      class="onyx-textarea__footer onyx-text--small"
-    >
-      <span v-if="errorMessages" class="onyx-textarea__error-message">
-        <span class="onyx-truncation-ellipsis">{{ errorMessages.shortMessage }}</span>
-
-        <OnyxInfoTooltip
-          v-if="errorMessages.longMessage"
-          class="onyx-textarea__message-tooltip"
-          color="danger"
-          position="bottom"
-          :label="t('showTooltip.error')"
-          :text="errorMessages.longMessage"
-        />
-      </span>
-      <span v-if="props.message" class="onyx-truncation-ellipsis">{{ props.message }}</span>
-      <OnyxInfoTooltip
-        v-if="props.messageTooltip"
-        class="onyx-textarea__message-tooltip"
-        position="bottom"
-        :text="props.messageTooltip"
-      />
-      <span v-if="shouldShowCounter" class="onyx-textarea__counter">
-        {{ value.length }}/{{ props.maxlength }}
-      </span>
-    </div>
+    </OnyxFormElement>
   </div>
 </template>
 
@@ -178,12 +124,6 @@ const handleInput = (event: Event) => {
 @use "../../styles/mixins/layers.scss";
 @use "../../styles/mixins/density.scss";
 @use "../../styles/mixins/input.scss";
-
-.onyx-use-optional:not(:has(.onyx-required-marker)) {
-  .onyx-textarea__optional {
-    display: inline-block;
-  }
-}
 
 .onyx-textarea,
 .onyx-textarea-skeleton {
@@ -240,17 +180,6 @@ const handleInput = (event: Event) => {
       $vertical-padding: var(--onyx-textarea-padding-vertical)
     );
 
-    &__header {
-      display: flex;
-      max-width: 100%;
-      width: 100%;
-    }
-
-    &__message-tooltip {
-      height: 1rem;
-      align-self: center;
-    }
-
     &__wrapper {
       padding: 0;
       height: unset;
@@ -274,10 +203,6 @@ const handleInput = (event: Event) => {
       &--no-resize {
         resize: none;
       }
-    }
-
-    .onyx-info-tooltip {
-      margin-left: var(--onyx-spacing-2xs);
     }
   }
 }